backend/health_ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10
pymupdf==1.23.8

# Use specific compatible versions
torch==2.0.1
transformers==4.31.0
sentence-transformers==2.2.2

numpy==1.24.3
python-dotenv==1.0.0
httpx==0.24.1